+++ pom.xml
@@ -66,6 +66,12 @@
 			<artifactId>spring-boot-starter-test</artifactId>
 			<scope>test</scope>
 		</dependency>
+		<dependency>
+			<groupId>org.awaitility</groupId>
+			<artifactId>awaitility</artifactId>
+			<!--			<version>4.2.0</version>-->
+			<scope>test</scope>
+		</dependency>
 	</dependencies>
 
 	<build>
