+++ src/main/java/com/genezeiniss/pos_transaction_processor/TransactionProcessorApplication.java
@@ -3,7 +3,9 @@ package com.genezeiniss.pos_transaction_processor;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.scheduling.annotation.EnableAsync;
 
+@EnableAsync
 @SpringBootApplication
 @EnableConfigurationProperties
 public class TransactionProcessorApplication {
