+++ src/main/java/com/genezeiniss/pos_transaction_processor/service/TransactionService.java
@@ -2,12 +2,9 @@ package com.genezeiniss.pos_transaction_processor.service;
 
 import com.genezeiniss.pos_transaction_processor.domain.Transaction;
 import com.genezeiniss.pos_transaction_processor.domain.TransactionMetadata;
-import com.genezeiniss.pos_transaction_processor.repository.TransactionMetadataRepository;
-import com.genezeiniss.pos_transaction_processor.repository.TransactionRepository;
 import com.genezeiniss.pos_transaction_processor.service.transaction_processor.TransactionProcessorFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Collections;
 import java.util.List;
@@ -18,10 +15,9 @@ import java.util.Optional;
 public class TransactionService {
 
     private final TransactionProcessorFactory transactionProcessorFactory;
-    private final TransactionRepository transactionRepository;
-    private final TransactionMetadataRepository transactionMetadataRepository;
+    private final PersistenceService persistenceService;
 
-    // todo: add logs
+    // todo: add logs - make sure that logs doesn't contain transaction metadata, which can include sensitive data
     public void processTransaction(Transaction transaction, List<TransactionMetadata> metadata) {
 
         var transactionProcessor = transactionProcessorFactory.getTransactionProcessor(transaction.getPaymentMethod());
@@ -29,15 +25,6 @@ public class TransactionService {
         transactionProcessor.validateTransactionOrException(transaction, Optional.ofNullable(metadata).orElse(Collections.emptyList()));
         transactionProcessor.processTransaction(transaction);
 
-        persistTransaction(transaction, metadata);
-    }
-
-    @Transactional
-    private void persistTransaction(Transaction transaction, List<TransactionMetadata> transactionMetadata) {
-        var trx = transactionRepository.create(transaction);
-        transactionMetadata.forEach(metadata -> {
-            metadata.setTransactionId(trx.getId());
-            transactionMetadataRepository.create(metadata);
-        });
+        persistenceService.persistTransaction(transaction, metadata);
     }
 }
