+++ src/test/java/com/genezeiniss/pos_transaction_processor/service/TransactionServiceTest.java
@@ -4,8 +4,6 @@ import com.genezeiniss.pos_transaction_processor.domain.TransactionMetadata;
 import com.genezeiniss.pos_transaction_processor.domain.enums.PaymentMethod;
 import com.genezeiniss.pos_transaction_processor.exception.ValidationException;
 import com.genezeiniss.pos_transaction_processor.fixture.TransactionFixture;
-import com.genezeiniss.pos_transaction_processor.repository.TransactionMetadataRepository;
-import com.genezeiniss.pos_transaction_processor.repository.TransactionRepository;
 import com.genezeiniss.pos_transaction_processor.service.transaction_processor.TransactionProcessor;
 import com.genezeiniss.pos_transaction_processor.service.transaction_processor.TransactionProcessorFactory;
 import org.junit.jupiter.api.DisplayName;
@@ -31,10 +29,7 @@ public class TransactionServiceTest {
     private TransactionProcessor transactionProcessor;
 
     @Mock
-    private TransactionRepository transactionRepository;
-
-    @Mock
-    private TransactionMetadataRepository transactionMetadataRepository;
+    private PersistenceService persistenceService;
 
     @InjectMocks
     private TransactionService transactionService;
@@ -54,6 +49,7 @@ public class TransactionServiceTest {
         verify(processorFactory).getTransactionProcessor(PaymentMethod.CASH);
         verify(transactionProcessor).validateTransactionOrException(transaction, metadata);
         verify(transactionProcessor).processTransaction(transaction);
+        verify(persistenceService).persistTransaction(transaction, metadata);
     }
 
     @Test
@@ -70,44 +66,6 @@ public class TransactionServiceTest {
                 () -> transactionService.processTransaction(transaction, metadata));
 
         verify(transactionProcessor, never()).processTransaction(transaction);
-        verifyNoInteractions(transactionRepository, transactionMetadataRepository);
-    }
-
-    @Test
-    @DisplayName("process transaction should persist transaction and metadata successfully")
-    void persistTransactionAndMetadata() {
-        var transaction = TransactionFixture.stubTransaction(PaymentMethod.CASH_ON_DELIVERY, 1.0);
-        var metadata = TransactionFixture.stubTransactionMetadata("courier", "yamato");
-        var metadataList = List.of(metadata);
-
-        when(processorFactory.getTransactionProcessor(PaymentMethod.CASH_ON_DELIVERY)).thenReturn(transactionProcessor);
-        doNothing().when(transactionProcessor).validateTransactionOrException(transaction, metadataList);
-        doNothing().when(transactionProcessor).processTransaction(transaction);
-
-        when(transactionRepository.create(transaction)).thenReturn(transaction);
-        when(transactionMetadataRepository.create(metadata)).thenReturn(metadata);
-
-        transactionService.processTransaction(transaction, metadataList);
-
-        verify(transactionRepository).create(transaction);
-        verify(transactionMetadataRepository).create(metadata);
-    }
-
-    @Test
-    @DisplayName("process transaction with empty metadata")
-    void persistTransactionOnly() {
-        var transaction = TransactionFixture.stubTransaction(PaymentMethod.CASH, 1.0);
-        List<TransactionMetadata> metadata = Collections.emptyList();
-
-        when(processorFactory.getTransactionProcessor(PaymentMethod.CASH)).thenReturn(transactionProcessor);
-        doNothing().when(transactionProcessor).validateTransactionOrException(transaction, metadata);
-        doNothing().when(transactionProcessor).processTransaction(transaction);
-
-        when(transactionRepository.create(transaction)).thenReturn(transaction);
-
-        transactionService.processTransaction(transaction, metadata);
-
-        verify(transactionRepository).create(transaction);
-        verify(transactionMetadataRepository, never()).create(any());
+        verifyNoInteractions(persistenceService);
     }
 }
